Restructure mailer crate and improve documentation
