+++ server/mailer/templates/password_changed.en.hbs
@@ -1,0 +1,14 @@
+<html lang="en">
+{{> header }}
+<body>
+<div class="container">
+    {{> banner }}
+
+    <p>Hi {{ name }}, </p>
+    <p>
+        Your password was changed. If this was not you, please reset your password using the reset password functionality.
+        If this was you, you do not need to do anything.
+    </p>
+</div>
+</body>
+</html>
