+++ server/mailer/templates/password_changed.nl.hbs
@@ -1,0 +1,14 @@
+<html lang="nl">
+{{> header }}
+<body>
+<div class="container">
+    {{> banner }}
+
+    <p>Hoi {{ name }}, </p>
+    <p>
+        Je wachtwoord is gewijzigd. Als jij dit niet was, verander je wachtwoord dan via de wachtwoord vergeten procedure.
+        Als jij dit wel was hoef je niets te doen.
+    </p>
+</div>
+</body>
+</html>
